# fix: Skip image decoding when in optimal format and not transforming.