### New Pull Request Checklist

* [x] I have read and understood the [CONTRIBUTING guide](https://github.com/rs/SDWebImage/blob/master/.github/CONTRIBUTING.md)
* [x] I have read the [Documentation](http://cocoadocs.org/docsets/SDWebImage/)
* [x] I have searched for a similar pull request in the [project](https://github.com/rs/SDWebImage/pulls) and found none

* [x] I have updated this branch with the latest master to avoid conflicts (via merge from master or rebase)
* [x] I have added the required tests to prove the fix/feature I am adding
* [x] I have updated the documentation (if necessary)
* [x] I have run the tests and they pass
* [x] I have run the lint and it passes (`pod lib lint`)

### Pull Request Description
#### Overview
This PR introduces an optimization to avoid unnecessary decode operations for image frames that are already in an optimal format for rendering. This improvement is particularly impactful for animated images with multiple frames, where decoding latency can significantly affect the viewing experience.

#### Key Changes
- Added format validation checks in `CGImageCreateDecoded:orientation:` to determine if decoding can be skipped
- Enhanced test coverage to verify optimization behavior
- Updated colorspace comparison in tests to check for equivalence rather than strict equality
- Fixed timing-based test flakiness in decode performance tests

#### Technical Details
The optimization checks for four key conditions before proceeding with decode:
1. Proper bit depth matching the target format
2. Compatible alpha channel configuration
3. Proper byte alignment for efficient GPU rendering
4. No orientation transformation needed

When all conditions are met, the original `CGImage` is returned without performing an expensive decode operation.

#### Performance Impact
- Reduces CPU overhead for images already in optimal format
- Particularly beneficial for animated images with many frames
- Improves memory usage by avoiding unnecessary image copies
- Enhanced viewing experience for animations due to reduced frame preparation time

#### Testing
Added comprehensive test coverage:
- Verification of decode skipping for optimal formats
- Proper decode behavior for images requiring transformation
- Colorspace handling across different platforms
- Byte alignment and pixel format validation

#### Manual Testing
The following logs show the `CGImageCreateDecoded` function performance before and after the change using an animated WebP at 828x466, illustrating the significant time associated with frame decoding.

**Before:**
```
2024-12-23 13:35:36.649408-0800 [25974:8907089] SDWebImage: Performed decode in 223.305ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312, size: 828x466
2024-12-23 13:35:36.881525-0800 [25974:8907089] SDWebImage: Performed decode in 222.237ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312, size: 828x466
2024-12-23 13:35:37.117319-0800 [25974:8907089] SDWebImage: Performed decode in 224.956ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312, size: 828x466
2024-12-23 13:35:37.353470-0800 [25974:8907089] SDWebImage: Performed decode in 227.337ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312, size: 828x466
2024-12-23 13:35:37.592824-0800 [25974:8907089] SDWebImage: Performed decode in 233.612ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312, size: 828x466
```

**After:**
```
2024-12-23 13:33:56.624034-0800 [24445:8901124] SDWebImage: Skipped decode in 0.003ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312
2024-12-23 13:33:56.642525-0800 [24445:8901131] SDWebImage: Skipped decode in 0.002ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312
2024-12-23 13:33:56.709962-0800 [24445:8901128] SDWebImage: Skipped decode in 0.002ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312
2024-12-23 13:33:56.793132-0800 [24445:8901131] SDWebImage: Skipped decode in 0.001ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312
2024-12-23 13:33:56.845217-0800 [24445:8901131] SDWebImage: Skipped decode in 0.002ms - hasAlpha: 0, orientation: 1, bpc: 8, bpp: 32, alignment: 3312
```