@dreampiggy 2024-12-24
This optimization's idea sounds OK.

The detailed check (`Whether the input CGImage can be directly render on iPhone screen without `CA::copy_image`) is hard. I even don't believe Apple's ImageIO and Core Animation team didn't change the condition between different iOS release.

So this is the reason why previous SD **Always do force decoding**

A idea, why not use `UIImage prepareForDisplay` for JPEG/HEIC ? Our default logic seems preferers to use Apple's UIKit top-level API as possible as it can.