@dreampiggy 2024-12-24
See more: https://jwilling.com/blog/debugging-core-animation-on-osx/