@skidder 2024-12-24
Thanks for sharing this post and the detailed explanation! Rather than trying to skip decoding, it sounds preferable to improve the performance of the force decode. I'll investigate the slow decoding of frames in my animated WebP test cases, and close out this PR for now. Thanks again!

> See more: https://jwilling.com/blog/debugging-core-animation-on-osx/
> 
> That `CA_COLOR_COPY` environment variable is useful. Does your new code breaks the optimization we previously want the `CGBitmapContext` to solve ?
> 
> Or can you use this variable to test whether Core Animation trigger the `CA::Render::copy_image` ? You need to render the new CGImage on the screen (via CALayer or UIImageView or NSImageView)