@dreampiggy 2024-12-25
> decoding of frames in my animated WebP test cases

Are you use `WebPCoder` (written by me and use libwebp) or ImageIO Apple based `AWebPCoder` ?

For `WebPCoder` we can use instruments and performent the issue. Maybe I think there are bugs that cause the extra unwanted `Force Decode and re-draw on bitmap context`.

Because `WebPCoder` decoded CGImage is allocated using `malloc`, not lazy, unlike ImageIO based `AWebPCoder` (which returned CGImage is lazy)