@skidder 2025-01-09
> Are you use `WebPCoder` (written by me and use libwebp) or ImageIO Apple based `AWebPCoder` ?

I was able to resolve these performance issues by switching from the Apple ImageIO coder to the SDWebImageWebPCoder. I incorrectly assumed that the Apple ImageIO WebP decoder would perform as well or better than the libwebp decoder, which just was not the case.

In my case, fixing the animated WebP decoding lag was as simple as adding the SDWebImageWebPCoder dependency to my app, and then switching from:

```
[[SDImageCodersManager sharedManager] addCoder:[SDImageAWebPCoder sharedCoder]];
```

To:
```
[[SDImageCodersManager sharedManager] addCoder:[SDImageWebPCoder sharedCoder]];
```

The [SDWebImage docs](https://github.com/SDWebImage/SDWebImage/wiki/Advanced-Usage#awebp-coder) hint at performance issues with ImageIO, but lack specifics. I think it'd be helpful to update them to describe situations where using libwebp might be best.

Thanks for your help!